+++ src/configs/ids.config.js
@@ -1,0 +1,82 @@
+const GENRE_IDS = {
+  "action": "1",
+  "adventure": "2",
+  "cars": "3",
+  "comedy": "4",
+  "dementia": "5",
+  "demons": "6",
+  "drama": "7",
+  "ecchi": "8",
+  "fantasy": "9",
+  "game": "10",
+  "harem": "11",
+  "historical": "12",
+  "horror": "13",
+  "isekai": "14",
+  "josei": "15",
+  "kids": "16",
+  "magic": "17",
+  "mahou-shoujo": "18",
+  "martial-arts": "19",
+  "mecha": "20",
+  "military": "21",
+  "music": "22",
+  "mystery": "23",
+  "parody": "24",
+  "police": "25",
+  "psychological": "26",
+  "romance": "27",
+  "samurai": "28",
+  "school": "29",
+  "sci-fi": "30",
+  "seinen": "31",
+  "shoujo": "32",
+  "shoujo-ai": "33",
+  "shounen": "34",
+  "shounen-ai": "35",
+  "slice-of-life": "36",
+  "space": "37",
+  "sports": "38",
+  "super-power": "39",
+  "supernatural": "40",
+  "thriller": "41",
+  "unknown": "42",
+  "vampire": "43"
+};
+
+const TYPE_IDS = {
+  "movie": "1",
+  "music": "2",
+  "ona": "3",
+  "ova": "4",
+  "special": "5",
+  "tv": "6"
+};
+
+const STATUS_IDS = {
+  "currently-airing": "1",
+  "finished-airing": "2",
+  "not-yet-aired": "3"
+};
+
+const RATING_IDS = {
+  "g": "1",
+  "pg": "2",
+  "pg-13": "3",
+  "r": "4",
+  "r+": "5",
+  "rx": "6"
+};
+
+const SORT_IDS = {
+  "default": "",
+  "latest-updated": "1",
+  "latest-added": "2",
+  "score": "3",
+  "name-az": "4",
+  "release-date": "5",
+  "most-viewed": "6",
+  "number_of_episodes": "7"
+};
+
+export { GENRE_IDS, TYPE_IDS, STATUS_IDS, RATING_IDS, SORT_IDS };